+++ src/main/java/org/terasology/launcher/gui/LauncherFrame.java
@@ -470,7 +470,7 @@ public final class LauncherFrame extends JFrame implements ActionListener {
         }
     }
 
-    private String escapeHtml(String text) {
+    private String escapeHtml(final String text) {
         return text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;")
             .replace("\"", "&quot;").replace("'", "&#x27;").replace("/", "&#x2F;");
     }