+++ src/main/java/org/terasology/launcher/version/GameJob.java
@@ -21,6 +21,9 @@ public enum GameJob {
     TerasologyStable("master", 15, 2, true, "STABLE",
         "infoHeader1_TerasologyStable", "settings_game_buildType_TerasologyStable"),
 
+    TerasologyLegacy("legacy", 1, 2, true, "STABLE",
+        "infoHeader1_TerasologyLegacy", "settings_game_buildType_TerasologyLegacy"),
+
     Terasology("develop", 245, 9, false, "NIGHTLY",
         "infoHeader1_Terasology", "settings_game_buildType_Terasology"),